Model with minimum concentration and corrected data
The corrected data from Randa used for model training
7 unknown parameters, including the modelling error variance and initial concentrations in roots and leaves + rate of uptake from matrix, rate of transfer from roots to leaves, rate of decay, and minimal concentration (Michaelis constant really).

No need to normalise concentration in matrix by matrix weight - t gets absorbed into the rate and is unidentifiable. The knowledge about the C_min also does not improve - suggesting that the model is not sensitive to it. Next step will be to remove it.

test_MCMC.py: testing the algorithm on the toy model that generates the ODE output and distorts it by measurement noise. The algorithm works ut the proposal must be very tight and it overestimates rate of decay. C_min posterior is uniform.

MCMC_iron.py: fits model to data for each individual experiment. The output of this is the MCMC chains corresponding to all model parameters.

read_MCMC_results.py: flattens the chains and visualises the results. The plan is to compare summary stats for the posteriors of all MCMC outputs to see which experiments can be described by the same model.

At the moment -- model of the uptake in the roots DOES not fit the data. This could be before the c_min. Needs checking!