Results for models with c_max and Michaelis constant
c_min is included as Michaelis constant first, incorporated as inhibiting factor in the uptake rate (fe_matrix - c_min)
Then it is included as maximal uptake capacity of roots induced by limited root area for uptake. In that model (c_max - fe_matrix) is used as inhibiting factor.
Both limiting constants are unidentifiable because the capacity is never reached in the experiments. The reamining model to do is without the uptake-limiting factor.